specops/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
django-debug-toolbar==3.4.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Pillow==9.1.0
sqlparse==0.4.2
